backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
idna==3.10
ifcopenshell==0.8.0
isodate==0.7.2
lark==1.2.2
numpy==2.1.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
shapely==2.0.6
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
